remove client side algolia related code/functionality/tests
