+++ app/frontend/packs/application.js
@@ -5,6 +5,7 @@ import 'styles/application.scss';
 
 import Rails from 'rails-ujs';
 
+import 'src/search/init';
 import 'src/lib/ui/patterns/currentLocation';
 import 'src/lib/ui/components/searchCheckbox';
 import 'src/addTitleToGoogleMapsIframe';
