--- app/frontend/src/search/__mocks__/query.js
@@ -1,3 +1,0 @@
-export const getQuery = jest.fn();
-
-export const getFilters = jest.fn();
