--- app/frontend/src/search/query.test.js
@@ -1,13 +1,0 @@
-import { getFilters, getQuery } from './query';
-
-describe('getFilters', () => {
-  test('returns a string', () => {
-    expect(typeof getFilters()).toBe('string');
-  });
-});
-
-describe('getQuery', () => {
-  test('returns an array', () => {
-    expect(typeof getQuery()).toBe('string');
-  });
-});
