--- app/frontend/src/search/ui/__mocks__/form.js
@@ -1,1 +1,0 @@
-export const enableSubmitButton = jest.fn();
