--- app/frontend/src/search/ui/input/__mocks__/keyword.js
@@ -1,1 +1,0 @@
-export const getKeyword = jest.fn();
