--- app/frontend/src/search/ui/pagination.test.js
@@ -1,7 +1,0 @@
-import { renderPagination } from './pagination';
-
-describe('renderPagination', () => {
-  test('returns a function', () => {
-    expect(typeof renderPagination).toBe('function');
-  });
-});
