+++ package.json
@@ -7,15 +7,11 @@
   "dependencies": {
     "@ministryofjustice/frontend": "^0.0.19-alpha",
     "@rails/webpacker": "4.2.2",
-    "algoliasearch": "~4.2.0",
     "axios": "^0.19.2",
     "clipboard": "^2.0.4",
-    "core-js": "^3.6.5",
     "govuk-frontend": "3.4.0",
-    "instantsearch.js": "^4.4.1",
     "jquery": "^3.5.0",
     "rails-ujs": "^5.2.3",
-    "regenerator-runtime": "^0.13.5",
     "rollbar": "^2.17.0"
   },
   "devDependencies": {
